refactor: move ns validation to where its used
