+++ pkg/okteto/namespace.go
@@ -15,10 +15,7 @@ package okteto
 
 import (
 	"context"
-	"fmt"
-	"regexp"
 
-	oktetoErrors "github.com/okteto/okteto/pkg/errors"
 	"github.com/okteto/okteto/pkg/types"
 	"github.com/shurcooL/graphql"
 )
@@ -131,23 +128,6 @@ func (c *namespaceClient) Delete(ctx context.Context, namespace string) error {
 		return err
 	}
 
-	return nil
-}
-
-func validateNamespace(namespace, object string) error {
-	if len(namespace) > MAX_ALLOWED_CHARS {
-		return oktetoErrors.UserError{
-			E:    fmt.Errorf("invalid %s name", object),
-			Hint: fmt.Sprintf("%s name must be shorter than 63 characters.", object),
-		}
-	}
-	nameValidationRegex := regexp.MustCompile("^[a-z0-9]([-a-z0-9]*[a-z0-9])?$")
-	if !nameValidationRegex.MatchString(namespace) {
-		return oktetoErrors.UserError{
-			E:    fmt.Errorf("invalid %s name", object),
-			Hint: fmt.Sprintf("%s name must consist of lower case alphanumeric characters or '-', and must start and end with an alphanumeric character", object),
-		}
-	}
 	return nil
 }
